added partitionner on producer, and some info on the partition on the consumer
